server/requirements.txt
fastapi
uvicorn
numpy
Pillow
opencv-python
tensorflow
